Fix warning comparison between types of different signedness

+++ src/mapping/AxialGeoMultiscaleMapping.cpp
@@ -108,9 +108,9 @@ void AxialGeoMultiscaleMapping::mapConsistent(const time::Sample &inData, Eigen:
                  "Unknown multiscale axis type.")
 
   // Check that the number of values for the input and output is right according to their dimensions
-  PRECICE_ASSERT((inputValues.size() / inDataDimensions == static_cast<int>(input()->vertices().size())),
+  PRECICE_ASSERT((inputValues.size() / static_cast<std::size_t>(inDataDimensions) == input()->vertices().size()),
                  inputValues.size(), inDataDimensions, input()->vertices().size());
-  PRECICE_ASSERT((outputValues.size() / outDataDimensions == static_cast<int>(output()->vertices().size())),
+  PRECICE_ASSERT((outputValues.size() / static_cast<std::size_t>(outDataDimensions) == output()->vertices().size()),
                  outputValues.size(), outDataDimensions, output()->vertices().size());
 
   // We currently don't support 1D data, so we need that the user specifies data of the same dimensions on both sides
